+++ src/Listener/UpdateCastorListener.php
@@ -3,7 +3,7 @@
 namespace Castor\Listener;
 
 use Castor\Console\Application;
-use Castor\PlatformUtil;
+use Castor\PlatformHelper;
 use JoliCode\PhpOsHelper\OsHelper;
 use Psr\Cache\CacheItemPoolInterface;
 use Psr\Log\LoggerInterface;
@@ -33,7 +33,7 @@ class UpdateCastorListener implements EventSubscriberInterface
         if (class_exists(\RepackedApplication::class)) {
             return;
         }
-        if (PlatformUtil::getEnv('DISABLE_VERSION_CHECK')) {
+        if (PlatformHelper::getEnv('DISABLE_VERSION_CHECK')) {
             return;
         }