+++ src/TaskDescriptorCollection.php
@@ -2,15 +2,13 @@
 
 namespace Castor;
 
-class TaskDescriptorCollection
+use Castor\Descriptor\TaskDescriptorCollection as BaseTaskDescriptorCollection;
+
+trigger_deprecation('castor', '0.15.0', 'The "%s" class is deprecated, use "%s" instead.', TaskDescriptorCollection::class, BaseTaskDescriptorCollection::class);
+
+/**
+ * @deprecated since Castor 0.15.0, use Castor\Descriptor\TaskDescriptorCollection instead
+ */
+class TaskDescriptorCollection extends BaseTaskDescriptorCollection
 {
-    /**
-     * @param TaskDescriptor[]        $taskDescriptors
-     * @param SymfonyTaskDescriptor[] $symfonyTaskDescriptors
-     */
-    public function __construct(
-        public readonly array $taskDescriptors,
-        public readonly array $symfonyTaskDescriptors,
-    ) {
-    }
 }