+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 * Add a way to merge an application `box.json` config file used by `castor:repack`command
 * Deprecate `Context::withPath()` in favor of `Context::withWorkingDirectory()`
 * Deprecate `path` argument in `capture()`, `exit_code()`, `run()`, `with()` in favor of `workingDirectory`
+* Deprecate `Castor\TaskDescriptorCollection` in favor of
+  `Castor\Descriptor\TaskDescriptorCollection`
 
 ## 0.14.0 (2024-03-08)